issue-16 Remove unnecessary code and extract polymorphic object structure
