+++ src/main/kotlin/pullrequestfactory/io/programs/impl/OpenPRProgram.kt
@@ -1,0 +1,5 @@
+package pullrequestfactory.io.programs.impl
+
+import pullrequestfactory.io.programs.Program
+
+interface OpenPRProgram : Program
